Split modeler so we can work with fewer merge conflicts
